policy-rag-api/requirements.txt
fastapi
uvicorn
langchain
ollama
redis
pydantic
python-dotenv
httpx
PyYAML
markdown
tiktoken
chroma
faiss-cpu
toml
aiofiles
pandas
